Hot fix data type streetFrontage

+++ src/main/java/com/majoapps/propertyfinder/data/entity/PropertyInformation.java
@@ -103,7 +103,7 @@ public class PropertyInformation {
     @Column(name="longitude")
     private Double longitude;
     @Column(name="street_frontage")
-    private Double streetFrontage;
+    private BigDecimal streetFrontage;
     @Column(name="legislation_url")
     private String legislationURL;
     @JsonIgnore
